Workflow wrapper; split WorkflowDetails into components w/ wrapper; update workflowdetails test
